Moved logic from controller to service
